Keep reward-excluded work out of MALIBU earning evidence
Constraint: Current main excludes relay-blind work from useful-work accrual; presentation must honor the same boundary.
Rejected: A direct rewards_excluded column reference | older mirrors do not contain that column.
Confidence: high
Scope-risk: narrow
Tested: Regression failed before fix; Docker race rewards suite passed after fix (41.596s); full coordinator tests, vet, isolated-cache lint, SPEC validation.
Not-tested: Production activation or payout execution; compute infrastructure remains unavailable.
